requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
appdirs==1.4.4
    # via requests-cache
attrs==21.4.0
    # via
    #   cattrs
    #   requests-cache
cattrs==1.10.0
    # via requests-cache
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via pynacl
charset-normalizer==2.0.12
    # via requests
click==8.0.4
    # via mkdocs
deprecated==1.2.13
    # via pygithub
ghp-import==2.0.2
    # via mkdocs
idna==3.3
    # via requests
importlib-metadata==4.11.1
    # via
    #   markdown
    #   mkdocs
jinja2==3.0.3
    # via
    #   mkdocs
    #   mkdocs-macros-plugin
    #   mkdocs-material
markdown==3.3.6
    # via
    #   mkdocs
    #   mkdocs-material
    #   pymdown-extensions
    #   pymdown-toc-ext
markupsafe==2.1.0
    # via jinja2
mergedeep==1.3.4
    # via mkdocs
mkdocs==1.2.3
    # via
    #   -r requirements.in
    #   mkdocs-macros-plugin
    #   mkdocs-material
    #   mkdocs-redirects
mkdocs-macros-plugin==0.6.4
    # via -r requirements.in
mkdocs-material==8.2.1
    # via -r requirements.in
mkdocs-material-extensions==1.0.3
    # via mkdocs-material
mkdocs-redirects==1.0.3
    # via -r requirements.in
packaging==21.3
    # via mkdocs
pycparser==2.21
    # via cffi
pygithub==1.55
    # via -r requirements.in
pygments==2.11.2
    # via mkdocs-material
pyjwt==2.3.0
    # via pygithub
pymdown-extensions==9.2
    # via mkdocs-material
pymdown-toc-ext==1.3.1
    # via -r requirements.in
pynacl==1.5.0
    # via pygithub
pyparsing==3.0.7
    # via packaging
python-dateutil==2.8.2
    # via
    #   ghp-import
    #   mkdocs-macros-plugin
python-dotenv==0.19.2
    # via -r requirements.in
python-frontmatter==1.0.0
    # via -r requirements.in
pyyaml==5.1
    # via
    #   mkdocs
    #   mkdocs-macros-plugin
    #   python-frontmatter
    #   pyyaml-env-tag
pyyaml-env-tag==0.1
    # via mkdocs
requests==2.27.1
    # via
    #   -r requirements.in
    #   pygithub
    #   requests-cache
requests-cache==0.9.2
    # via -r requirements.in
six==1.16.0
    # via
    #   mkdocs-redirects
    #   python-dateutil
    #   url-normalize
termcolor==1.1.0
    # via mkdocs-macros-plugin
url-normalize==1.4.3
    # via requests-cache
urllib3==1.26.8
    # via
    #   requests
    #   requests-cache
watchdog==2.1.6
    # via mkdocs
wrapt==1.13.3
    # via deprecated
zipp==3.7.0
    # via importlib-metadata
